Use deltaTime in grid movement
Because we used fixed update, we don't have to multiply by deltaTime to prevent it from being framerate independent. However, we may want to change the physics time step (I plan on changing it to 60 per sec later probably), and this allows us to change that without having to adjust speeds.