Disable Auto Sync Transforms
It seems to only be important when raycasting to a moving collider which we don't do. I didn't notice any difference in our case, but according to Unity 'This will eliminate spikes due to TransformChangeDispatch and Physics scene updates from Physics queries.' They are also scheduling it to be off by default soon.