Create a game state manager
This helps us to implement pausing in a proper way. Scripts can access the game state from the gameState member of the game controller. We still need to implement activity for the suspend state, use C# properties instead of get and set functions because I didn't do that for whatever reason, and eventually clean up unused states after implementing all modes that we want.