Allow asynchronously loading Game scene after main menu is shown, but disable by default
Since the main menu doesn't do so much, loading in the background shouldn't be an issue. Doing so gives a percent change of about -70% which is great, except that the cost of calling the async load will slow the main menu loading time to worse than the Game scene synchronously, ouch. We work around that by ensuring a frame of the main menu is rendered before we start the async load, but even then calling loadSceneAsync will eat frames causing inputs to be potentially dropped, so we stick with the slow synchronous loading by default as a consequence. Maybe Unity will make asynchronous scene loading viable for something other than animated load screens somewhere down the road.