Fix and improve Directions.rotate one more time
It says the rotation is relative, but it is only relative along one axis. Making it entirely relative broke direction arrows, as they have a rotated parent that isn't the game map. The solution is to allow us to specify a transform to rotate relative to so that everything can be rotated about the game map. Unfortunately we need to keep a map reference in maptile as a result, and I didn't measure if performance noticibly changed, but it is good to be accurate beyond what is currently considered necessary.