Allow the player to reset on failure early
This has a bug in which the auto reset timer isn't stopped since our timer doesn't support that yet, but this is the extremely rare instance where I'm going to commit this anyway since I plan on adding that timer feature as the next thing I do. We also only support one map ping ring now.