Use events over polling for some inputs
For events that are not per frame and very rare (which is everything that isn't cursor movement) we want to prefer events since the garbage generated will be negligible and it saves us some frame time. Reduces time for PlayerController update to a tenth of what it was (though the high cpu usage from mouse movement was unable to be fixed and probably needs fixed from Unity's end).