Make sure walls are entirely above floor
As far as I understand, since they had equal center positions, there was no clear way to do front to back sorting for walls and the floor, which could break batching, so make sure the wall center is always above the floor. Didn't verify if this made a difference though.